updateServer now has onUpdateServer (same with the client version) The thinking here is that it was possible to override updateServer and not call the inherited method which checks for loaded images, updates children, makes videos play etc. This lead to a bunch of hard-to-figure out bugs. So now we enforce that updateServer is properly run through before letting client apps respond to updates. NOTE: if you are extending a class that also uses onUpdateServer or onUpdateClient, you'll still need to call that class's onUpdateServer. DO NOT call updateServer from an onUpdateServer method, it will create an infinite loop.

+++ example/media_viewer/src/app/mediaviewer_app.cpp
@@ -5,8 +5,6 @@
 #include <Poco/Path.h>
 
 #include <cinder/Rand.h> 
-#include <cinder/app/RendererGl.h>
-#include <cinder/app/RendererGl.h>
 
 #include <ds/app/environment.h>
 #include <ds/debug/logger.h>
@@ -172,9 +170,9 @@ void MediaViewer::keyDown(ci::app::KeyEvent event){
 	// Shows all enabled sprites with a label for class type
 	} else if(event.getCode() == KeyEvent::KEY_f){
 
-		const int numRoots = mEngine.getRootCount();
+		const size_t numRoots = mEngine.getRootCount();
 		int numPlacemats = 0;
-		for(int i = 0; i < numRoots - 1; i++){
+		for(size_t i = 0; i < numRoots - 1; i++){
 			mEngine.getRootSprite(i).forEachChild([this](ds::ui::Sprite& sprite){
 				if(sprite.isEnabled()){
 					sprite.setTransparent(false);
@@ -200,8 +198,8 @@ void MediaViewer::keyDown(ci::app::KeyEvent event){
 		}
 		mStreamer = new ds::ui::GstVideo(mEngine);
 		mStreamer->startStream(mGlobals.getSettingsLayout().getText("streaming:pipeline", 0, ""),
-							  mGlobals.getSettingsLayout().getInt("streaming:width", 0, 640), 
-							  mGlobals.getSettingsLayout().getInt("streaming:height", 0, 480));
+							 (float) mGlobals.getSettingsLayout().getInt("streaming:width", 0, 640), 
+							 (float) mGlobals.getSettingsLayout().getInt("streaming:height", 0, 480));
 		mStreamerParent->addChildPtr(mStreamer);
 		mStreamer->enable(true);
 		mStreamer->enableMultiTouch(ds::ui::MULTITOUCH_CAN_POSITION | ds::ui::MULTITOUCH_CAN_SCALE);
